Create dedicated package for strings benchmarks
