+++ benchmarks/src/main/scala/zio/redis/strings/AppendBenchmarks.scala
@@ -1,9 +1,10 @@
-package zio.redis
+package zio.redis.strings
 
 import java.util.concurrent.TimeUnit
 
 import org.openjdk.jmh.annotations._
 
+import zio.redis._
 import zio.ZIO
 
 @State(Scope.Thread)
@@ -12,7 +13,7 @@ import zio.ZIO
 @Measurement(iterations = 15)
 @Warmup(iterations = 15)
 @Fork(2)
-class GetBenchmarks extends BenchmarkRuntime {
+class AppendBenchmarks extends BenchmarkRuntime {
 
   @Param(Array("500"))
   var count: Int = _
@@ -31,23 +32,23 @@ class GetBenchmarks extends BenchmarkRuntime {
     import _root_.laserdisc.{ all => cmd, _ }
     import cats.instances.list._
     import cats.syntax.foldable._
-    unsafeRun[LaserDiscClient](c => items.traverse_(i => c.send(cmd.get[String](Key.unsafeFrom(i)))))
+    unsafeRun[LaserDiscClient](c => items.traverse_(i => c.send(cmd.append(Key.unsafeFrom(i), i))))
   }
 
   @Benchmark
   def rediculous(): Unit = {
     import cats.implicits._
     import io.chrisdavenport.rediculous._
-    unsafeRun[RediculousClient](c => items.traverse_(i => RedisCommands.get[RedisIO](i).run(c)))
+    unsafeRun[RediculousClient](c => items.traverse_(i => RedisCommands.append[RedisIO](i, i).run(c)))
   }
 
   @Benchmark
   def redis4cats(): Unit = {
     import cats.instances.list._
     import cats.syntax.foldable._
-    unsafeRun[Redis4CatsClient[String]](c => items.traverse_(i => c.get(i)))
+    unsafeRun[Redis4CatsClient[String]](c => items.traverse_(i => c.append(i, i)))
   }
 
   @Benchmark
-  def zio(): Unit = zioUnsafeRun(ZIO.foreach_(items)(get[String, String]))
+  def zio(): Unit = zioUnsafeRun(ZIO.foreach_(items)(i => append(i, i)))
 }
